Folders structure refactor (models, views,..)
