+++ src/main/java/fr/lirmm/fairness/assessment/models/requestparams/ParamTest.java
@@ -1,4 +1,4 @@
-package fr.lirmm.fairness.assessment.utils.requestparams;
+package fr.lirmm.fairness.assessment.models.requestparams;
 
 
 public interface ParamTest {
