+++ src/main/java/fr/lirmm/fairness/assessment/models/requestparams/RequestParamValidator.java
@@ -1,4 +1,4 @@
-package fr.lirmm.fairness.assessment.utils.requestparams;
+package fr.lirmm.fairness.assessment.models.requestparams;
 
 import javax.servlet.http.HttpServletRequest;
