+++ src/main/java/fr/lirmm/fairness/assessment/models/requestparams/exceptions/ApikeyException.java
@@ -1,0 +1,7 @@
+package fr.lirmm.fairness.assessment.models.requestparams.exceptions;
+
+public class ApikeyException extends  NoValidParam{
+    public ApikeyException(String message) {
+        super(message);
+    }
+}
