+++ src/main/java/fr/lirmm/fairness/assessment/principles/criterion/AbstractPrincipleCriterion.java
@@ -10,8 +10,8 @@ import java.util.stream.Collectors;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import fr.lirmm.fairness.assessment.models.Configuration;
-import fr.lirmm.fairness.assessment.utils.QuestionResult;
-import fr.lirmm.fairness.assessment.utils.Result;
+import fr.lirmm.fairness.assessment.models.results.QuestionResult;
+import fr.lirmm.fairness.assessment.models.results.Result;
 import fr.lirmm.fairness.assessment.principles.AbstractScoredEntity;
 import fr.lirmm.fairness.assessment.principles.criterion.question.AbstractCriterionQuestion;
 import org.json.JSONException;
