+++ src/main/java/fr/lirmm/fairness/assessment/principles/criterion/question/Tester.java
@@ -2,7 +2,7 @@ package fr.lirmm.fairness.assessment.principles.criterion.question;
 
 import fr.lirmm.fairness.assessment.models.Ontology;
 import fr.lirmm.fairness.assessment.models.Property;
-import fr.lirmm.fairness.assessment.utils.QuestionResult;
+import fr.lirmm.fairness.assessment.models.results.QuestionResult;
 
 import java.util.List;
 import java.util.Map;
