--- src/main/java/fr/lirmm/fairness/assessment/utils/Environments.java
@@ -1,5 +1,0 @@
-package fr.lirmm.fairness.assessment.utils;
-
-public enum Environments {
-    DEV,PROD
-}
