--- src/main/java/fr/lirmm/fairness/assessment/utils/requestparams/exceptions/ApikeyException.java
@@ -1,7 +1,0 @@
-package fr.lirmm.fairness.assessment.utils.requestparams.exceptions;
-
-public class ApikeyException extends  Exception{
-    public ApikeyException(String message) {
-        super(message);
-    }
-}
